+++ Packages/com.unity.inputsystem/InputSystem/InputManager.cs
@@ -1310,6 +1310,9 @@ namespace UnityEngine.InputSystem
 
             // Let listeners know.
             DelegateHelpers.InvokeCallbacksSafe(ref m_DeviceChangeListeners, device, InputDeviceChange.Removed, "onDeviceChange");
+
+            // Try setting next device of same type as current
+            InputSystem.GetDevice(device.GetType())?.MakeCurrent();
         }
 
         public void FlushDisconnectedDevices()
@@ -1683,17 +1686,30 @@ namespace UnityEngine.InputSystem
             }
         }
 
+        private unsafe void QueueEvent(InputEvent* eventPtr)
+        {
+            // If we're currently in OnUpdate(), the m_InputEventStream will be open. In that case,
+            // append events directly to that buffer and do *NOT* go into native.
+            if (m_InputEventStream.isOpen)
+            {
+                m_InputEventStream.Write(eventPtr);
+                return;
+            }
+
+            // Don't bother keeping the data on the managed side. Just stuff the raw data directly
+            // into the native buffers. This also means this method is thread-safe.
+            m_Runtime.QueueEvent(eventPtr);
+        }
+
         public unsafe void QueueEvent(InputEventPtr ptr)
         {
-            m_Runtime.QueueEvent(ptr.data);
+            QueueEvent(ptr.data);
         }
 
         public unsafe void QueueEvent<TEvent>(ref TEvent inputEvent)
             where TEvent : struct, IInputEventTypeInfo
         {
-            // Don't bother keeping the data on the managed side. Just stuff the raw data directly
-            // into the native buffers. This also means this method is thread-safe.
-            m_Runtime.QueueEvent((InputEvent*)UnsafeUtility.AddressOf(ref inputEvent));
+            QueueEvent((InputEvent*)UnsafeUtility.AddressOf(ref inputEvent));
         }
 
         public void Update()
@@ -1979,6 +1995,7 @@ namespace UnityEngine.InputSystem
         private bool m_NativeBeforeUpdateHooked;
         private bool m_HaveDevicesWithStateCallbackReceivers;
         private bool m_HasFocus;
+        private InputEventStream m_InputEventStream;
 
         // We want to sync devices when the editor comes back into focus. Unfortunately, there's no
         // callback for this so we have to poll this state.
@@ -2783,6 +2800,8 @@ namespace UnityEngine.InputSystem
             //       execution (and we're not sure where it's coming from).
             Profiler.BeginSample("InputUpdate");
 
+            Debug.Assert(!m_InputEventStream.isOpen, "Already have an event buffer set! Was OnUpdate() called recursively?");
+
             // Restore devices before checking update mask. See InputSystem.RunInitialUpdate().
             RestoreDevicesAfterDomainReloadIfNecessary();
 
@@ -2874,38 +2893,30 @@ namespace UnityEngine.InputSystem
                 return;
             }
 
-            var currentEventReadPtr =
-                (InputEvent*)NativeArrayUnsafeUtility.GetUnsafeBufferPointerWithoutChecks(eventBuffer.data);
-            var remainingEventCount = eventBuffer.eventCount;
             var processingStartTime = Stopwatch.GetTimestamp();
-
-            // When timeslicing events or in before-render updates, we may be leaving events in the buffer
-            // for later processing. We do this by compacting the event buffer and moving events down such
-            // that the events we leave in the buffer form one contiguous chunk of memory at the beginning
-            // of the buffer.
-            var currentEventWritePtr = currentEventReadPtr;
-            var numEventsRetainedInBuffer = 0;
-
             var totalEventLag = 0.0;
 
             #if UNITY_EDITOR
             var isPlaying = m_Runtime.isInPlayMode;
             #endif
+            
+            m_InputEventStream = new InputEventStream(ref eventBuffer);
 
             // Handle events.
-            while (remainingEventCount > 0)
+            while (m_InputEventStream.remainingEventCount > 0)
             {
                 InputDevice device = null;
+                var currentEventReadPtr = m_InputEventStream.currentEventPtr;
 
-                Debug.Assert(!currentEventReadPtr->handled);
+                Debug.Assert(!currentEventReadPtr->handled, "Event in buffer is already marked as handled");
 
                 // In before render updates, we only take state events and only those for devices
                 // that have before render updates enabled.
                 if (updateType == InputUpdateType.BeforeRender)
                 {
-                    while (remainingEventCount > 0)
+                    while (m_InputEventStream.remainingEventCount > 0)
                     {
-                        Debug.Assert(!currentEventReadPtr->handled);
+                        Debug.Assert(!currentEventReadPtr->handled, "Iterated to event in buffer that is already marked as handled");
 
                         device = TryGetDeviceById(currentEventReadPtr->deviceId);
                         if (device != null && device.updateBeforeRender &&
@@ -2913,11 +2924,10 @@ namespace UnityEngine.InputSystem
                              currentEventReadPtr->type == DeltaStateEvent.Type))
                             break;
 
-                        eventBuffer.AdvanceToNextEvent(ref currentEventReadPtr, ref currentEventWritePtr,
-                            ref numEventsRetainedInBuffer, ref remainingEventCount, leaveEventInBuffer: true);
+                        currentEventReadPtr = m_InputEventStream.Advance(leaveEventInBuffer: true);
                     }
                 }
-                if (remainingEventCount == 0)
+                if (m_InputEventStream.remainingEventCount == 0)
                     break;
 
                 var currentEventTimeInternal = currentEventReadPtr->internalTime;
@@ -3032,8 +3042,7 @@ namespace UnityEngine.InputSystem
 
                 if (skipEvent)
                 {
-                    eventBuffer.AdvanceToNextEvent(ref currentEventReadPtr, ref currentEventWritePtr,
-                        ref numEventsRetainedInBuffer, ref remainingEventCount, leaveEventInBuffer: leaveInBuffer);
+                    m_InputEventStream.Advance(leaveEventInBuffer: leaveInBuffer);
                     continue;
                 }
 
@@ -3161,8 +3170,7 @@ namespace UnityEngine.InputSystem
                         break;
                 }
 
-                eventBuffer.AdvanceToNextEvent(ref currentEventReadPtr, ref currentEventWritePtr,
-                    ref numEventsRetainedInBuffer, ref remainingEventCount, leaveEventInBuffer: false);
+                m_InputEventStream.Advance(leaveEventInBuffer: false);
             }
 
             m_Metrics.totalEventProcessingTime += ((double)(Stopwatch.GetTimestamp() - processingStartTime)) / Stopwatch.Frequency;
@@ -3170,26 +3178,10 @@ namespace UnityEngine.InputSystem
 
             // Remember how much data we retained so that we don't count it against the next
             // batch of events that we receive.
-            InputUpdate.s_LastUpdateRetainedEventCount = (uint)numEventsRetainedInBuffer;
-            InputUpdate.s_LastUpdateRetainedEventBytes = (uint)((byte*)currentEventWritePtr -
-                (byte*)NativeArrayUnsafeUtility
-                    .GetUnsafeBufferPointerWithoutChecks(eventBuffer
-                    .data));
+            InputUpdate.s_LastUpdateRetainedEventCount = (uint)m_InputEventStream.numEventsRetainedInBuffer;
+            InputUpdate.s_LastUpdateRetainedEventBytes = m_InputEventStream.numBytesRetainedInBuffer;
 
-            // Update event buffer. If we have retained events, update event count
-            // and buffer size. If not, just reset.
-            if (numEventsRetainedInBuffer > 0)
-            {
-                var bufferPtr = NativeArrayUnsafeUtility.GetUnsafeBufferPointerWithoutChecks(eventBuffer.data);
-                Debug.Assert((byte*)currentEventWritePtr > (byte*)bufferPtr);
-                var newBufferSize = (byte*)currentEventWritePtr - (byte*)bufferPtr;
-                eventBuffer = new InputEventBuffer((InputEvent*)bufferPtr, numEventsRetainedInBuffer, (int)newBufferSize,
-                    (int)eventBuffer.capacityInBytes);
-            }
-            else
-            {
-                eventBuffer.Reset();
-            }
+            m_InputEventStream.Close(ref eventBuffer);
 
             if (gameIsPlayingAndHasFocus)
                 ProcessStateChangeMonitorTimeouts();