+++ Packages/com.unity.inputsystem/Tests/TestFixture/InputTestRuntime.cs
@@ -70,6 +70,7 @@ namespace UnityEngine.InputSystem
 
                     m_EventCount = buffer.eventCount;
                     m_EventWritePosition = (int)buffer.sizeInBytes;
+
                     if (NativeArrayUnsafeUtility.GetUnsafeBufferPointerWithoutChecks(buffer.data) !=
                         NativeArrayUnsafeUtility.GetUnsafeBufferPointerWithoutChecks(m_EventBuffer))
                         m_EventBuffer = buffer.data;
@@ -383,12 +384,14 @@ namespace UnityEngine.InputSystem
 
         public int eventCount => m_EventCount;
 
+        internal const int kDefaultEventBufferSize = 1024 * 512;
+
         private bool m_HasFocus = true;
         private int m_NextDeviceId = 1;
         private int m_NextEventId = 1;
         internal int m_EventCount;
         private int m_EventWritePosition;
-        private NativeArray<byte> m_EventBuffer = new NativeArray<byte>(1024 * 1024, Allocator.Persistent);
+        private NativeArray<byte> m_EventBuffer = new NativeArray<byte>(kDefaultEventBufferSize, Allocator.Persistent);
         private List<PairedUser> m_UserPairings;
         private List<KeyValuePair<int, string>> m_NewDeviceDiscoveries;
         private List<KeyValuePair<int, DeviceCommandCallback>> m_DeviceCommandCallbacks;